Update/pl update individual custom field

See merge request mmlabel/mm-printing-backend!77

+++ internal/aurora/service/production_plan/edit.go
@@ -3,6 +3,7 @@ package production_plan
 import (
 	"context"
 	"fmt"
+	"strings"
 	"time"
 
 	"mmlabel.gitlab.com/mm-printing-backend/internal/aurora/model"
@@ -94,11 +95,7 @@ func (c *productionPlanService) EditProductionPlan(ctx context.Context, opt *Edi
 		plan.Thumbnail = cockroach.String(opt.Thumbnail)
 		plan.Note = cockroach.String(opt.Note)
 		plan.Status = opt.Status
-		currentStage, ok := enum.ProductionPlanStatusSage[plan.Status]
-		if !ok {
-			return fmt.Errorf("invalid plan status: %v", plan.Status)
-		}
-		plan.CurrentStage = currentStage // only current stage can view the production plan
+		//plan.CurrentStage = opt.CurrentStage // TODO: have to update this field
 
 		if err := c.productionPlanRepo.Update(ctx2, plan.ProductionPlan); err != nil {
 			return fmt.Errorf("update production plan failed: %w", err)
@@ -115,6 +112,14 @@ func (c *productionPlanService) EditProductionPlan(ctx context.Context, opt *Edi
 			}
 		}
 		for _, field := range deletedCustomFields {
+			// if field.key contains sale_survey then ignore delete
+
+			if strings.Contains(field.Field, "sale_survey") {
+				// Todo: this hard code is not good, but we have to do it,
+				// because there are 2 forms: production plan and sale survey ( Thu thap thong tin sale)
+				continue
+			}
+
 			if err := c.customFieldRepo.Delete(ctx, field.ID); err != nil {
 				return fmt.Errorf("delete custom field for production plan failed: %w", err)
 			}
@@ -124,6 +129,70 @@ func (c *productionPlanService) EditProductionPlan(ctx context.Context, opt *Edi
 	})
 	if execTx != nil {
 		return execTx
+	}
+
+	return nil
+}
+
+func (c *productionPlanService) UpdateCustomFields(ctx context.Context, productionPlanID string, fieldValues []*CustomField) error {
+	plan, err := c.productionPlanRepo.FindByID(ctx, productionPlanID)
+	if err != nil {
+		return err
+	}
+	if !plan.ProductionPlan.Editable() {
+		return fmt.Errorf("không thể chỉnh sửa kế hoạch đã được đưa vào sản xuất")
+	}
+	errTx := cockroach.ExecInTx(ctx, func(ctx2 context.Context) error {
+
+		customFields, err := c.customFieldRepo.Search(ctx, &repository.SearchCustomFieldsOpts{
+			EntityId:   plan.ID,
+			EntityType: enum.CustomFieldTypeProductionPlan,
+			Limit:      1000,
+			Offset:     0,
+		})
+
+		if err != nil {
+			return fmt.Errorf("search custom fields failed: %w", err)
+		}
+
+		customFieldMap := generic.ToMap(customFields, func(f *repository.CustomFieldData) string {
+			return f.Field
+		})
+
+		newCustomFields := make([]*model.CustomField, 0)
+		updatedCustomFields := make([]*model.CustomField, 0)
+		for _, field := range fieldValues {
+			if _, ok := customFieldMap[field.Field]; ok {
+				customFieldMap[field.Field].Value = field.Value
+				updatedCustomFields = append(updatedCustomFields, customFieldMap[field.Field].CustomField)
+				delete(customFieldMap, field.Field)
+			} else {
+				newCustomFields = append(newCustomFields, &model.CustomField{
+					ID:         idutil.ULIDNow(),
+					EntityID:   plan.ID,
+					EntityType: enum.CustomFieldTypeProductionPlan,
+					Field:      field.Field,
+					Value:      field.Value,
+				})
+			}
+		}
+		for _, field := range updatedCustomFields {
+			if err := c.customFieldRepo.Update(ctx, field); err != nil {
+				return fmt.Errorf("update custom field for production plan failed: %w", err)
+			}
+		}
+		for _, field := range newCustomFields {
+			if err := c.customFieldRepo.Insert(ctx, field); err != nil {
+				return fmt.Errorf("create custom field for production plan failed: %w", err)
+			}
+		}
+
+		return nil
+
+	})
+
+	if errTx != nil {
+		return fmt.Errorf("update custom fields failed: %w", errTx)
 	}
 
 	return nil
